pyproject.toml
[project]
name = "gpt-sovits"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "av>=14.4.0",
    "chardet>=5.2.0",
    "cn2an>=0.5.23",
    "ctranslate2>=4.6.0",
    "fast-langdetect>=0.3.2",
    "fastapi[standard]>=0.115.12",
    "ffmpeg-python>=0.2.0",
    "funasr==1.0.27",
    "g2p-en>=2.1.0",
    "g2pk2>=0.0.3",
    "gradio>=5.31.0",
    "huggingface-hub>=0.32.0",
    "jieba>=0.42.1",
    "jieba-fast>=0.53",
    "ko-pron>=1.3",
    "librosa==0.10.2",
    "matplotlib>=3.10.3",
    "modelscope==1.10.0",
    "numba>=0.61.2",
    "numpy<=2.0",
    "onnxruntime>=1.22.0",
    "onnxruntime-gpu>=1.22.0",
    "opencc==1.1.1",
    "peft>=0.15.2",
    "psutil>=7.0.0",
    "pydantic>=2.11.5",
    "pydub>=0.25.1",
    "pyopenjtalk>=0.4.1",
    "pypinyin>=0.54.0",
    "pytorch-lightning>=2.5.1.post0",
    "pyyaml>=6.0.2",
    "rotary-embedding-torch>=0.8.6",
    "scipy>=1.15.3",
    "sentencepiece>=0.2.0",
    "split-lang>=2.1.0",
    "tensorboard>=2.19.0",
    "tojyutping>=3.2.0",
    "tokenizers>=0.21.1",
    "torchaudio>=2.7.0",
    "torchmetrics>=1.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.3",
    "uvicorn>=0.34.2",
    "wordsegment>=1.3.1",
    "x-transformers>=2.3.5",
]
